Update ShipIt to use objects instead of static classes
Summary:
In order to eventually support configs being pulled from configerator, we will need to be able to support configs that are generated -- by definition non-static.

The first step in this process is to convert all of the static `classname<FBShipItConfig` variables into `FBShipItConfig` instances. Then, later diffs will slowly cut-over to use non-`static` versions of all of the `FBShipItConfig` functions.

Reviewed By: yns88

Differential Revision: D15635277

fbshipit-source-id: c4cd5fb9cf0e1de0b62afb1e02e13451d717b7d1