Be open and transparent (protected -> public)
Summary: There are a few functions in FBShipItConfig that are supported in configerator but we don't support in the conversion script yet. This adds support for them, and, because they were marked `protected`, they also had to be converted to public, which necessitated a lot of boring command-line fu to make the typchecker behave.

Reviewed By: yns88

Differential Revision: D15814553

fbshipit-source-id: 2f023d16b9bab313d2ec949fb906a5475e7d8530