Implement comment markers as a standard option
Summary:
First we define 3 new overloadable functions that together form the options for commenting and uncommenting lines for export:

## getCommentMarker()
This is nullable and defaults to null, comments are only filtered if a non-null value is set here in an overloaded function. This will typically be '//' for most languages

## getCommentLine()
This defaults to "oss-disable" and is used to determine which lines to comment out as they're being exported

## getUncommentLine()
This defaults to "oss-enable" and is used to determine which lines to comment out as they're being imported

I went ahead and modified the `FBWWWBase` to use these new features, In another diff I will enable the migration script to import these functions to configerator and then update `FBShipItConfigeratorConfig` to use them.

Reviewed By: yns88

Differential Revision: D15946301

fbshipit-source-id: 3bcbfdc22109a083302312cf8771a104e7983692