initial ShipIt config for OSS itchef cookbooks
Summary:
We have a repo, `https://github.com/facebook/IT-CPE` that contains many open-sourced Chef cookbooks, that currently have to be synced manually.

Get ShipIt working, but initially don't ship any cookbook yet. Once we've verified that ShipIt works we'll add `cpe_helpers` cookbook (split out of `cpe_utils`) and get it shipped out.

Reason for publishing to `itchef/` instead of to the root of the repo: we have existing files there and want ShipIt to fully manage this brand-new directory to avoid the potential of falling out of sync.

Reviewed By: joshuamiller01

Differential Revision: D15979983

fbshipit-source-id: 8d7c095bee9d7ac79ff433f1ee6e0c871b4f72d2