Prevent the meerkat job from pulling in hack plugins
Summary:
The meerkat job needs to be updated to not add `FBShipItconfigeratorConfig`-descendant classes to the class map.

The `FBShipItconfigeratorConfig`-descendant classes are used as hack plugins, so they will be pulled in by the configerator configs automatically, they don't need to be part of the classmap

This causes problems in the meerkat job because they shouldn't be instantiated and crash the job when it tries to filter TagIt classes a few lines later.

Reviewed By: yns88

Differential Revision: D16007328

fbshipit-source-id: bac973b9be2e41040d47b6eaf26c3ddf9d7c2673