Pass local vars by inout instead of by ref in ShipItShellCommand::runOnceSynchronously
Summary: We're banning references in Hack. This change changes the calls to use inout which will invoke an implicit wrapper in HHVM that will translate from inout to reffy. This change should be okay b/c the inputs shouldn't be accessible if the call is interrupted pre-maturely (e.g. throws).

Reviewed By: periodic1236

Differential Revision: D16099718

fbshipit-source-id: 3c24cd7ecce98f62ff3c8e65745f18b5dcfcf624