Add hack plugins to generated configs
Summary:
Since I'm about to convert ~58 configs to use plugins it would be nice to have an ergonomic way to do the conversion.

This patch adds an optional, overridable function to the base `FBShipIt` that will signal the migration script to __ignore__ its decline reasoning logic and convert the config no matter what, and use the returned value of this function as the hack plugins name during migration.

Reviewed By: cdelahousse

Differential Revision: D16096200

fbshipit-source-id: 380c7c11472dd24770c875806155c7e266670d48