Upgrade hacktest, drop support for HHVM 3.30 (#76)
Summary:
We need a newer version of hacktest because of a codemod that was done on expect().

However, the new version of hacktest does not support HHVM 3.30, so we have to drop support for that. 4.8 appears to be a semi-LTS version, so add that to the build matrix instead.
Pull Request resolved: https://github.com/facebook/fbshipit/pull/76

Reviewed By: zpao

Differential Revision: D16228244

Pulled By: yns88

fbshipit-source-id: b1572bbf246cbd5b9885c0885933a7f6b07fee8e