Filter out ROCm backend paths in ShipIt
Summary: This commit modifies the Glow ShipIt config to exclude paths with the prefix `lib/Backends/ROCm`. The goal is to allow development of the Glow ROCm backend within fbcode without cloning the entire `fbcode/glow/glow` directory as `glow-nnpi` did, and to keep the work private for now. Open sourcing this work at a later date should be as easy as reverting this diff.

Reviewed By: arunm-git

Differential Revision: D16274296

fbshipit-source-id: 232ece1872d6a3ba8c7d320ce646b4fc5e04aa30