Enable FBCodeBuilder configs to use configerator
Summary:
I've duplicated the `FBCodeBuilderTrait` to be able to slowly transition these configs. Once the transition is complete, the old one will be deleted and the new one will be renamed.

I also went ahead and translated the first config (`mvfst`) to demonstrate how this will work. The configerator config can be found in D16006812.

Reviewed By: yns88

Differential Revision: D16008356

fbshipit-source-id: 28d7ef93a64253fc9f303c3ae95a5a2ed9358947