Clean up leftover __destruct references in www.
Summary:
`__destruct` has done nothing in the runtime for a long time now.
Delete the one remaining `__destruct` function; it was already marked `__OptionalDestruct`, but it was definitely doing nothing.
Stop trying to unstub `Curl::__destruct` in GraphCurlStubHack::unstubCurl. It doesn't exist any more, and we're not stubbing it in the first place in GraphCurlStubHack::stubCurl.

Reviewed By: yns88

Differential Revision: D16531858

fbshipit-source-id: c0bc841b974d5ebfbb555c9b71ec1bea50d0e82d